% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp1qc_solver.R
\name{binsearchtol}
\alias{binsearchtol}
\title{binary search with arbitrary resolution}
\usage{
binsearchtol(fun, tol = 0.01, range, ...)
}
\arguments{
\item{fun}{a monotonic function over which we search (passed to \code{\link{binsearch}})}

\item{tol}{resolution of the grid over which to search for \code{target} (see \code{\link{binsearch}})}

\item{range}{a range over which to search for the input to \code{fun}}

\item{...}{passed to \code{\link{binsearch}}}
}
\description{
The \code{\link{binsearch}} function in the \code{gtools} package searches only over integers. This is a wrapper for \code{\link{binsearch}} that also allows searching over a grid of points with distance \code{tol} between them. A target must also be entered (see \code{\link{binsearch}}).
}
\examples{
 # best solution here is at x0,
 # which we can find with increasing precision
 x0 <- 10.241
 binsearchtol( function(x) x-x0, target=0, range=c(0,2*x0) , tol=1.00)
 binsearchtol( function(x) x-x0, target=0, range=c(0,2*x0) , tol=0.10)
 binsearchtol( function(x) x-x0, target=0, range=c(0,2*x0) , tol=0.05)

}
\seealso{
\code{\link{binsearch}}
}
