% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp1qc_solver.R
\name{simDiagBasis}
\alias{simDiagBasis}
\title{Simultaneously diagonalize two symmetric matrices (1 being positive definite)}
\usage{
simDiagBasis(M1, M2, eigenM1 = NULL, eigenM2 = NULL, tol,
  return_diags = FALSE)
}
\arguments{
\item{M1}{a positive definite symmetric matrix}

\item{M2}{a symmetric matrix}

\item{tol}{used for error checks}

\item{return_diags}{should Q^(-1); Q'M1Q; and Q'BQ be returned.
Resulting matrix Q gives  Q'M1Q=I (identity); Q'BQ = a diagonal matrix.
Q is *not* orthonormal or even symmetric, but it is invertible.
#########}
}
\description{
Simultaneously diagonalize two symmetric matrices (1 being positive definite)
}
